+++ src/controller/MiColaboracionController.java
@@ -35,19 +35,28 @@ public class MiColaboracionController {
 		this.miColaboracionDao = colaboracionDao;
 	}
 	
-	@RequestMapping("/listar")
-	public String listaColaboracion(HttpSession session, Model model){
-		Usuario usuario = (Usuario) session.getAttribute("usuario");
-		model.addAttribute("colaboracionesOferta", miColaboracionDao.getMisColaboracionesOferta(usuario.getUsuario()));
-		model.addAttribute("colaboracionesPeticion", miColaboracionDao.getMisColaboracionesPeticion(usuario.getUsuario()));
-		return "miColaboracion/listar";
+	@RequestMapping("/listar/{usuario}")
+	public String listaColaboracion(HttpSession session, Model model, @PathVariable String usuario){
+		Usuario u = (Usuario) session.getAttribute("usuario");
+		if(u != null && (u.getUsuario().equals(usuario) || u.getRol().name().equals("ADMIN"))){
+			model.addAttribute("colaboracionesOferta", miColaboracionDao.getMisColaboracionesOferta(usuario));
+			model.addAttribute("colaboracionesPeticion", miColaboracionDao.getMisColaboracionesPeticion(usuario));
+			return "miColaboracion/listar";
+		}else{
+			return "error/error";
+		}
 	}
 	
 
 	@RequestMapping(value="/add")
 	public String addColaboracion(HttpSession session, Model model){
-		model.addAttribute("miColaboracion", new MiColaboracion());
-		return "miColaboracion/add";
+		Usuario u = (Usuario) session.getAttribute("usuario");
+		if(u != null){
+			model.addAttribute("miColaboracion", new MiColaboracion());
+			return "miColaboracion/add";
+		}else{
+			return "error/error";
+		}
 	}
 
 	@RequestMapping(value="/add", method=RequestMethod.POST)
@@ -58,10 +67,15 @@ public class MiColaboracionController {
 		return "redirect:listar.html";
 	}
 	
-	@RequestMapping(value="/update/{id_colaboracion}", method = RequestMethod.GET)
-	public String editColaboracion(HttpSession session, Model model, @PathVariable int id_colaboracion){
+	@RequestMapping(value="/update/{usuario}/{id_colaboracion}", method = RequestMethod.GET)
+	public String editColaboracion(HttpSession session, Model model, @PathVariable int id_colaboracion, @PathVariable String usuario){
+		Usuario u = (Usuario) session.getAttribute("usuario");
+		if(u != null && (u.getUsuario().equals(usuario) || u.getRol().name().equals("ADMIN"))){	
 			model.addAttribute("miColaboracion", miColaboracionDao.getMiColaboracion(id_colaboracion));
-			return "miColaboracion/update";	
+			return "miColaboracion/update";
+		}else{
+			return "error/error";
+		}
 	}
 	
 	@RequestMapping(value="/update/{id_colaboracion}", method = RequestMethod.POST)
@@ -73,10 +87,15 @@ public class MiColaboracionController {
 		return "redirect:../listar.html";
 	}
 	
-	@RequestMapping(value="/delete/{id_colaboracion}")
-	public String processDelete(HttpSession session, @PathVariable int id_colaboracion){
-		miColaboracionDao.deleteMiColaboracion(id_colaboracion);
-		return "redirect:../listar.html";
+	@RequestMapping(value="/delete/{usuario}/{id_colaboracion}")
+	public String processDelete(HttpSession session, @PathVariable int id_colaboracion, @PathVariable String usuario){
+		Usuario u = (Usuario) session.getAttribute("usuario");
+		if(u != null && (u.getUsuario().equals(usuario) || u.getRol().name().equals("ADMIN"))){
+			miColaboracionDao.deleteMiColaboracion(id_colaboracion);
+			return "redirect:../listar.html";
+		}else{
+			return "error/error";
+		}
 	}
 	
 	/*formateo de fechas	 */
